Big merge of all development changes over last year into main

+++ lib/common_widgets/image/circular_network_image/circular_network_image.dart
@@ -1,11 +1,11 @@
-import 'package:fingerfunke_app/common_widgets/image/network_placeholder_asset_image/network_placeholder_asset_image.dart';
+import '../network_placeholder_asset_image/network_placeholder_asset_image.dart';
 import 'package:flutter/material.dart';
-import 'package:fingerfunke_app/utils/type_aliases.dart';
+import '../../../utils/type_aliases.dart';
 
 class CircularNetworkImage extends StatelessWidget {
   final Link imageUrl;
   final AssetImage placeholder;
-  final int? diameter;
+  final double? diameter;
   const CircularNetworkImage(this.imageUrl, this.placeholder,
       {Key? key, this.diameter})
       : super(key: key);
