+++ lib/common_widgets/image/network_placeholder_asset_image/network_placeholder_asset_image.dart
@@ -1,12 +1,12 @@
-import 'package:fingerfunke_app/common_widgets/image/network_placeholder_image.dart/network_placeholder_image.dart';
-import 'package:fingerfunke_app/utils/type_aliases.dart';
+import '../network_placeholder_image.dart/network_placeholder_image.dart';
+import '../../../utils/type_aliases.dart';
 import 'package:flutter/material.dart';
 
 class NetworkPlaceholderAssetImage extends StatelessWidget {
   final Link imageUrl;
   final AssetImage placeholder;
-  final int? height;
-  final int? width;
+  final double? height;
+  final double? width;
   final BoxFit? fit;
   const NetworkPlaceholderAssetImage(this.imageUrl, this.placeholder,
       {Key? key, this.height, this.width, this.fit})
