+++ lib/common_widgets/image/user_image/user_image.dart
@@ -1,10 +1,11 @@
-import 'package:fingerfunke_app/common_widgets/image/circular_network_image/circular_network_image.dart';
 import 'package:flutter/material.dart';
-import 'package:fingerfunke_app/utils/type_aliases.dart';
+
+import '../../../utils/type_aliases.dart';
+import '../circular_network_image/circular_network_image.dart';
 
 class UserImage extends StatelessWidget {
   final Link? imageUrl;
-  final int? diameter;
+  final double? diameter;
   final AssetImage placeholder =
       const AssetImage("assets/img/placeholder/profile_avatar_placeholder.png");
   const UserImage(this.imageUrl, {Key? key, this.diameter}) : super(key: key);
